Added support to not store security statements in query and file history. Refactored code base.
